[Fix] API code fix
