+++ SKHUAZ/SKHUAZ/Network/Router/UserRouter/UserRouter.swift
@@ -1,5 +1,5 @@
 //
-//  LogInRouter.swift
+//  SignUpRouter.swift
 //  SKHUAZ
 //
 //  Created by 문인호 on 2023/09/25.
@@ -9,12 +9,15 @@ import Foundation
 
 import Alamofire
 
-enum LogInRouter {
+enum UserRouter {
     case LogIn (request: LogInRequest)
     case nicknameCheck (nickname: String )
     case LogOut(token: String)
     case editProfile(request: EditProfileRequest, token: String)
     case signOut(token: String)
+    case SignUp (request: SignUpRequest)
+    case emailAuth (request: emailSendRequest)
+    case emailCode (request: emailCodeRequest)
 }
 
 struct LogInRequest: Encodable {
@@ -35,7 +38,30 @@ struct EditProfileRequest: Encodable {
         case doubleMajor = "double_major"
     }
 }
-extension LogInRouter: BaseTargetType {
+
+struct emailSendRequest: Encodable {
+    let email: String
+}
+
+struct emailCodeRequest: Encodable {
+    let email: String
+    let code: String
+}
+
+struct SignUpRequest: Encodable {
+    let email: String
+    let password: String
+    let nickname: String
+    let semester: String
+    let graduate: Bool
+    let major1: String
+    let major2: String
+    let department: Bool
+    let major_minor: Bool
+    let double_major: Bool
+}
+
+extension UserRouter: BaseTargetType {
     
     var baseURL: String { return Config.baseURL }
     
@@ -50,6 +76,12 @@ extension LogInRouter: BaseTargetType {
         case .signOut:
             return .delete
         case .editProfile:
+            return .post
+        case .SignUp:
+            return .post
+        case .emailAuth:
+            return .post
+        case .emailCode:
             return .post
         }
     }
@@ -66,6 +98,12 @@ extension LogInRouter: BaseTargetType {
             return "/user/delete"
         case .editProfile:
             return "/user/update-information"
+        case .SignUp:
+            return "user/join"
+        case .emailAuth:
+            return "email/send"
+        case .emailCode:
+            return "email/verify/code"
         }
     }
     
@@ -81,22 +119,35 @@ extension LogInRouter: BaseTargetType {
             return .none
         case let .editProfile(request, _):
             return .body(request)
+        case let .SignUp(request):
+            return .body(request)
+        case let .emailAuth(request):
+            return .body(request)
+        case let .emailCode(request):
+            return .body(request)
         }
-    }
-    var headers : HTTPHeaders?{
-        switch self{
-        case .LogIn:
-            return nil
-        case .nicknameCheck:
-            return nil
-        case .LogOut(let token):
-            return ["Authorization": "Bearer \(token)"]
-        case .signOut(let token):
-            return ["Authorization": "Bearer \(token)"]
-        case .editProfile(_, let token):
-            return ["Authorization": "Bearer \(token)"]
-        default:
-            return nil
+        
+        var headers : HTTPHeaders?{
+            switch self{
+            case .LogIn:
+                return nil
+            case .nicknameCheck:
+                return nil
+            case .LogOut(let token):
+                return ["Authorization": "Bearer \(token)"]
+            case .signOut(let token):
+                return ["Authorization": "Bearer \(token)"]
+            case .editProfile(_, let token):
+                return ["Authorization": "Bearer \(token)"]
+            case .SignUp:
+                return nil
+            case .emailAuth:
+                return nil
+            case .emailCode:
+                return nil
+            default:
+                return nil
+            }
         }
     }
 }
